+++ src/interpolation.cc
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2015-2018,2020
+ *    Copyright (c) 2015-2018,2020,2026
  *      SMASH Team
  *
  *    GNU General Public License (GPLv3 or later)
@@ -11,15 +11,27 @@
 
 namespace smash {
 
-InterpolateDataSpline::InterpolateDataSpline(const std::vector<double>& x,
-                                             const std::vector<double>& y) {
+InterpolateDataSpline::InterpolateDataSpline(
+    const std::vector<double>& x, const std::vector<double>& y,
+    const ExtrapolationType extrapolation_type)
+    : extrapolation_type_{extrapolation_type} {
+  switch (extrapolation_type_) {
+    case ExtrapolationType::None:
+    case ExtrapolationType::Zero:
+    case ExtrapolationType::Constant:
+      break;
+    default:
+      throw std::invalid_argument(
+          "The provided extrapolation type is not supported. Valid types are "
+          "'None', 'Zero', and 'Constant'.");
+  }
   const auto N = x.size();
   if (y.size() != N) {
-    throw std::runtime_error(
-        "Need two vectors of equal length for interpolation.");
+    throw std::invalid_argument(
+        "The interpolation requires two vectors of equal length.");
   }
   if (N < 3) {
-    throw std::runtime_error(
+    throw std::invalid_argument(
         "Need at least 3 data points for cubic spline interpolation.");
   }
   const auto p = generate_sort_permutation(
@@ -43,12 +55,18 @@ InterpolateDataSpline::~InterpolateDataSpline() {
 }
 
 double InterpolateDataSpline::operator()(double xi) const {
-  // constant extrapolation
-  if (xi < first_x_) {
-    return first_y_;
-  }
-  if (xi > last_x_) {
-    return last_y_;
+  if (xi < first_x_ || xi > last_x_) {
+    if (extrapolation_type_ == ExtrapolationType::None) {
+      std::stringstream error_msg{
+          "InterpolateDataSpline only accepts x values within the range of the "
+          "underlying data when an extrapolation type is not specified."};
+      error_msg << "\nx value " << xi << " is out of bounds.";
+      throw std::out_of_range(error_msg.str());
+    } else if (extrapolation_type_ == ExtrapolationType::Zero) {
+      return 0.;
+    } else if (extrapolation_type_ == ExtrapolationType::Constant) {
+      return (xi < first_x_) ? first_y_ : last_y_;
+    }
   }
   // cubic spline interpolation
   return gsl_spline_eval(spline_, xi, acc_);