+++ CHANGELOG.md
@@ -24,8 +24,8 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 * Leading hadrons are now selected not only based on their proximity in momentum space to the string endpoints, but also on whether the corresponding endpoint originates from a valence quark of the incoming beams.
   This reduces the number of leading hadrons at higher energies, where previously two leading hadrons were selected per string.
 * `CrossSections::probability_transit_high` has been replaced by `CrossSections::interpolation_at_sqrts`.
-* New key `Ignore_Unformed` in `Output: Thermodynamics` to exclude unformed (and preformed) particles from the density
-evaluation. This is useful e.g. in afterburner calculations.
+* New key `Ignore_Unformed` in `Output: Thermodynamics` to exclude unformed (and preformed) particles from the density evaluation. This is useful e.g. in afterburner calculations.
+* The path to tabulations can now be passed with the command line argument `-t to/path/tabulations`. However, if the input files for particles/decay modes or SMASH version change, the tabulations will be recalculated in the indicated folder. This is useful to make several runs share the same tabulations folder, even if these are in different output directories, avoiding creating several sets of possibly identical files.
 
 ### Changed
 * ⚠️ To build SMASH tuning compilation for a different architecture than the `native` one, the `SMASH_ARCH_FLAG` CMake variable must now be used (e.g. passing `-DSMASH_ARCH_FLAG=x86-64` to CMake). Using `CMAKE_CXX_FLAGS` and `CMAKE_C_FLAGS` or setting the environment `CFLAGS` and `CXXFLAGS` variables will not have the desired effect!
@@ -33,9 +33,13 @@ evaluation. This is useful e.g. in afterburner calculations.
 * Minimum supported ROOT version is now version `6.20`.
 * Minimum supported HepMC3 version is now version `3.2.6`.
 * The `make install` target has been standardized to create an enclosing version-specific folder inside which the standard `bin`, `lib`, `include` and `share` directories are created (previously, the version specific folder was created inside the standard ones).
+* Leading hadrons are now selected not only based on their proximity in momentum space to the string endpoints, but also on whether the corresponding endpoint originates from a valence quark of the incoming beams. This reduces the number of leading hadrons at higher energies, where previously two leading hadrons were selected per string.
+* `CrossSections::probability_transit_high` has been replaced by `CrossSections::interpolation_at_sqrts`.
+* The interpolations `InterpolateDataLinear`, `InterpolateDataSpline`, and `InterpolateData2DSpline` are restricted to not extrapolate by default and will in this case now throw if out-of-bounds values are tried to access.
+  Extrapolation can be enabled by giving an `ExtrapolationType` as an argument during construction.
 
 ### Removed
-* CMake utility function `add_compiler_flags_if_supported` has been removed as it sets flags globally
+* CMake utility function `add_compiler_flags_if_supported` has been removed as it sets flags globally.
 
 ## SMASH-3.3
 Date: 2025-12-03